requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
coverage==5.2
dj-database-url==0.5.0
Django==3.0.8
django-oidc-provider==0.7.0
djangorestframework==3.11.0
future==0.18.2
idna==2.10
psycopg2-binary==2.8.5
pycryptodomex==3.9.8
pyjwkest==1.4.2
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
